chore: added adapter for sensor values, updated sensor drawing #PATRON2020-247

+++ app/src/main/java/com/intive/patronage/smarthome/feature/dashboard/model/MapPosition.kt
@@ -2,4 +2,4 @@ package com.intive.patronage.smarthome.feature.dashboard.model
 
 import com.squareup.moshi.Json
 
-data class MapPosition(@Json(name = "x") val x: Int, @Json(name = "y") val y: Int)
+data class MapPosition(@Json(name = "x") val x: Float, @Json(name = "y") val y: Float)
